requirements.txt
open3d==0.18.0
torch_scatter==2.1.2
torchmetrics==1.3.2
numpy
open_clip_torch
opencv-python
setuptools
networkx
scipy
scikit-learn
matplotlib
natsort

